Add QuizModelStub class
